@@ -3,36 +3,21 @@
 	"version": "0.0.1",
 	"scripts": {
 		"dev": "vite dev",
-		"build": "vite build && npm run package",
-		"preview": "vite preview",
-		"package": "svelte-kit sync && svelte-package && publint",
-		"prepublishOnly": "npm run package"
-	},
-	"exports": {
-		".": {
-			"types": "./dist/index.d.ts",
-			"svelte": "./dist/index.js"
-		}
+		"build": "vite build",
+		"preview": "vite preview"
 	},
 	"files": ["dist", "!dist/**/*.test.*", "!dist/**/*.spec.*"],
-	"peerDependencies": {
-		"svelte": "^4.0.0"
-	},
 	"devDependencies": {
 		"@biomejs/biome": "1.9.4",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/package": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"publint": "^0.1.9",
-		"svelte": "^4.2.7",
-		"typescript": "^5.3.2",
-		"vite": "^5.0.11"
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.15.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"normalize.css": "^8.0.1",
+		"svelte": "^5.16.0",
+		"typescript": "^5.7.2",
+		"vite": "^6.0.6"
 	},
 	"svelte": "./dist/index.js",
 	"types": "./dist/index.d.ts",
-	"type": "module",
-	"dependencies": {
-		"normalize.css": "^8.0.1"
-	}
+	"type": "module"
 }
