@@ -2,6 +2,14 @@
 import "normalize.css";
 
 import "$lib/app.css";
+
+/**
+ * @typedef {Object} Props
+ * @property {import('svelte').Snippet} [children]
+ */
+
+/** @type {Props} */
+let { children } = $props();
 </script>
 
-<slot />
+{@render children?.()}
